Hopefully more useful error messages
Improves the formatting of error messages somewhat (this is completely
subjective, but I like them better).

Also, adds JSONPath as a dependency to print context of the error.
Instead of printing the entire object for every error, just print the
parent of the node with the error.

If this is still too much data (in the case where a large parent and all
of its children are being printed), line 107 can be commented out. Then
only the node with the error will be printed.